+++ app/models/magistrate.js
@@ -12,22 +12,6 @@ export default Profile.extend({
   // Attributes
 
   /**
-   * Nombre completo del cargo.
-   *
-   * @property cargoNombreCompleto
-   * @type String
-   */
-  cargoNombreCompleto: attr('string'),
-
-  /**
-   * Nombre corto del cargo.
-   *
-   * @property cargoNombreCorto
-   * @type String
-   */
-  cargoNombreCorto: attr('string'),
-
-  /**
    * Profesión.
    *
    * @property profesion
@@ -60,14 +44,6 @@ export default Profile.extend({
   desempenio: attr('string'),
 
   /**
-   * Historial politico.
-   *
-   * @property historialPolitico
-   * @type String
-   */
-  historialPolitico: attr('string'),
-
-  /**
    * Experiencia profesional.
    *
    * @property experienciaProfesional
@@ -76,15 +52,7 @@ export default Profile.extend({
   experienciaProfesional: attr('string'),
 
   /**
-   * Experiencia en derechos humanos.
-   *
-   * @property experienciaEnDH
-   * @type String
-   */
-  experienciaEnDH: attr('string'),
-
-  /**
-   * Información general. Campos para mostrar en información general. 
+   * Información general. Campos para mostrar en información general. Son serializados después.
    *
    * @property informacionGeneral
    * @type informacion-general
@@ -100,12 +68,36 @@ export default Profile.extend({
   recuadros: attr('frente-a-frente'),
 
   /**
-   * Campos a evaluar en la modularidad frente a frente.
+   * Campos a evaluar en la modularidad frente a frente. Son serializados después.
    *
    * @property frenteAFrente
    * @type frente-a-frente
    */
   frenteAFrente: attr('frente-a-frente'),
+
+  /**
+   * Institución
+   *
+   * @property institucion
+   * @type String
+   */
+  institucion: attr('string'),
+
+  /**
+   * Plan de trabajo
+   *
+   * @property planTrabajo
+   * @type String
+   */
+  planTrabajo: attr('string'),
+
+  /**
+   * Proyección
+   *
+   * @property proyeccion
+   * @type String
+   */
+  proyeccion: attr('string'),
 
   // Computed Properties