+++ tests/unit/models/commission-deputies-test.js
@@ -7,7 +7,7 @@ module('Unit | Model | commission deputies', function(hooks) {
   // Replace this with your real tests.
   test('it exists', function(assert) {
     let store = this.owner.lookup('service:store');
-    let model = store.createRecord('commission-deputies', {});
+    let model = store.createRecord('commission-deputie', {});
     assert.ok(model);
   });
 });