+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added main initializer.
 - Added ember's addons ember-metrics, ember-router-scroll.
+- Added YUIDoc at spreadsheet service.
 
 ## Changed
 - Actions in index controller.